Serve /goverment/feed
This is an Atom feed currently served by Whitehall. It is generated
using the same method as the organisations feed. When we migrated that
feed we identified two functional changes that will also apply here:

- The `id` of the feed will change from
`tag:www.gov.uk,2005:DocumentCollection/231443` to a URL.
- We can no longer provide long form feed content as Rummager doesn't
let us obtain that. We fall back to providing a title and a summary.