Make taxon browsing tests less brittle
We want to assert that sections and links are present. The titles of these
elements are only partially controlled in this app now (the titles are now
contained in the govuk_document_types gem).  By being less picky about the text
that we match on, we can hopefully reduce the number of changes we need to make
in the future.