Render courts and HMCTS tribunals in this app
Courts and HMCTS tribunals are those pages accessed via
https://www.gov.uk/courts-tribunals.  They're still rendered by Whitehall at
present.

There's not a lot for us to do in order to render these pages, as most of the
work was already done by the org pages migration.

We don't show latest documents, as there are specialist finders set up for
things like Employment Tribunal decisions.  These pages link to those where
appropriate.

We don't show the FOI section, as this is the current state on Whitehall.

We don't show the corporate information section as this is the current state on
Whitehall.

We don't show the "Read more" link at the bottom of the "What we do" section
because we render the content from the about page in this section already.
Showing the link would just link to the content that the user has just read,
which would seem unhelpful.

I'm using the same style of test as the [taxon pages](https://github.com/alphagov/collections/blob/master/test/integration/taxon_browsing_test.rb).
I've put them in a separate test file from the other org pages, partly because
the tests are in a different style.

I propose to move the org pages tests to this style at some point too.

https://trello.com/c/IRoHCQ3N/143-render-courts-and-tribunals-pages-from-collections