Translate content purpose into document types
Iterating which document types belong to which supertype is very slow if it
means republishing content and getting it all reindexed by rummager.  We are
circumventing this by ignoring the `content_purpose_supergroup` properties of
content items and rummager results, and instead translating them prior to
performing queries.

This has already been done in [finder frontend](https://github.com/alphagov/finder-frontend/pull/619),
and this change is to ensure that collections is in line with these changes.