Test that GovukDocumentTypes returns sane things
We have devolved the responsibility of translating content purposes into lists
of document types to the govuk_document_types gem.

In the tests which exercise this path, we've stubbed the response from this to
ensure that we reduce the brittleness of the tests as much as possible.

So that we can retain confidence that the gem returns reasonable responses for
each of the groups (and that the groups still exist in the gem), we've added a
few tests to assert how we expect the gem to behave.