Fix duplicate tracking events for See All News
Duplicate tracking events were being sent when a user clicked on See All News and Communications in this topic.
This is because the link was wrapped by a parent with data-module=track-click while it also had that attribute itself.

This is a work-around. Ideally we should modify the document list component so it adds the track click module itself rather than having to wrap the component when we include it.