Refactor citizen landing page to improve performance
This commit refactors the citizen landing page to remove the need to create new `Topic` instances, instead using `ContentItem` data. For `other_taxons` in `BrexitTaxonsPresenter`, the `ContentItem` of `/` can be used directly to create `ContentItem`s and remove the need for any additional calls to the Content Store.

Solo: @karlbaker02