Get consultation closing dates from rummager
This means we no longer have to do lookups to content store for the extra
information which should speed up page rendering significantly as well as
making the render process more robust as we need to make fewer api requests.

This was added to Whitehall in https://github.com/alphagov/whitehall/pull/4611 (and
all consultations have been reindexed), so all the information should be present.