Catch incorrect organisation api requests
/api/organisations is a prefix route so can get some odd things sent to it
which at present fall through to the taxons catch all at the bottom of routes.rb
resulting in a 500 error.

It's best if we can avoid 500s for incorrect user requests, so let's catch them
where we can.  This is a naive attempt which rejects things that are under the api
namespace, but haven't matched the simple route of /api/organisations/<org-name>.

The proc response is enough to convey a 404 error.