Cache search queries by default
These parameters are generally fixed which makes the calls to rummager fairly
cacheable.  We don't mind if the results are a little stale as we'll otherwise
fall back to the mirrors on a 5xx error which can be up to 24 hours old.

In practice these won't get hit repeatedly over a short period except when
Fastly doesn't have a cached response (for example requests with query
strings).

Some of these pages make multiple requests to render a single page, so caching
successful requests should add some resilience to the process.  That is,
if a page needs to make four requests to search and the third one fails, then
the first two should already be cached so will not need to be made again.