The homepage section about government departments has two adjacent links that both link to the same page, but have different text.
**How does this affect users**
Confusing for screen readers and confusing for cognitive issues
WCAG 2.4.4 (A)

To mitigate against this, we should amend both links and the page they point to, so that links anchor to the relevant section.

To do this we first need to add a mechanism to link to relevant sections.

This work adds the `organisation_type` id to each section heading.