Use component title for Courts headers
The h1 for [other organisations](https://www.gov.uk/government/organisations/ministry-of-housing-communities-and-local-government)
is small because we can be more reliant on other aspects of the page,
such as the organisation logo, featured items and the "latest from
<department>" sections to inform users where they are.

Courts and tribunals pages don't have these things, so it makes much more
sense to have a more visually significant page header.  Fortunately we can
use the title component to do this.