Add headings between screenshots
A long wall of screenshots without any space in between them,
and with no context other than the alt text, was not particularly
useful. I've taken the alt text and used this as a h3 value above
each screenshot so that it's clearer what the significance of each
screenshot is, and to make the screenshots linkable via an anchor
link.