The BrexitLandingPagePresenter now uses BrexitSections
- Add unit tests
- This is the Brexit Landing Page, so the content item is brexit
by definition
- Removed the filter on show_sections; this is a conditional that
checks if any of the supergroup links has any content associated
with it. This makes sense in a regular Taxon page because some
taxons do not have any content for a supergroup and displaying
the link would cause clutter. For the Brexit Landing Page this
is not the case and checking for it would cause unnecessary
overhead and complexity.
- changed map->each
- removed some superfluous memoization