Fix chevron jagged border issue
Because the chevron SVG was overlaying the link text div, the two borders were interfering with each other. At times, the two borders would appear next to each other and cause a jagged appearance. This switches to using an SVG that starts exactly at the point the chevron triangle starts, to avoid the overlap, and also sets attributes on the stroke to stop it scaling.