Create A and B show for landing page
This commit creates two variants of the landing page, A and you guessed
it... B. Variant B is fairly bare but will be updated with the new
landing page designs in a later PR.

The action knows which view to render via a http header passed in, if
this isn't present the default rendered view will be A. When a variant
is rendered a html meta tag is also set for the response, this tells GA which
version of the page was used.

link to chevron-extend-hover header has been removed due to it being
redundant as the code for this resides in govuk-publishing-components.