Add URLs to steps in HowTo schema
The absence of URLs are currently showing as warnings in the structured
data testing tool. We can resolve this one quickly as steps already have
a deterministic ID based on the title.

I've updated the component to use the [same method to generate IDs](https://github.com/alphagov/govuk_publishing_components/pull/1152).
It's nearly the same already, just it doesn't collapse multiple dashes.