Assign images to each step of the HowTo schema
We should be using unique images for each step in the howTo schema.  As we do
not, we surmise that we're not eligible for our preferred rich result
type in Google.  Kate's kindly provided me with some images to use.

We start numbering at 1 to match the component.

If the step has a logic "and" or "or" then we choose the appropriate image
and do not increment the step counter (because we want to use the same numbering
style as the component).

Otherwise, we choose the appropriate image from the list.

We don't have any more than 12 steps at the moment.  In the case of
having more, then we just use the placeholder text.