Make i18n translation keys clearer
The keys in the i18n translation yaml files normally refer to the text.

However, as the content here is subject to a lot of change, this actually makes it more difficult to see what translations are being used for, and they can become very out of date easily. This renames some of the translations so they refer to where on the page they are being used. These are locations which are unlikely to change, e.g: page title, meta descriptions etc are unlikely to be changed to be rendered elsewhere on the page.