Rescue 403 errors from content store
At the moment collections is throwing a 500, when it receives an error code it doesn't recognise
from content-store. That means that they are shown a "Problem has occurred" page.

We need to handle this error properly so that in the case of a 403, the user will be properly routed
to signon and asked to login if there is any form of access limiting on the content item (e.g. host on draft
stack, access limited to organisation)