Render people in a layout similar to Whitehall
We're currently looking into whether we could migrate people and roles
out of Whitehall.

While investigating, I discovered that we already have the document type
set up in collections, and we already have quite a lot of the
information required in the Content Store.

I've never done any frontend work before, but I wanted to have a go so
this changes the layout of the page in collections to match the current
layout in Whitehall.