Fix topic browse localisations not working
The presentation of the 'topic list' section on the Brexit Landing Page
is based on a set of 'Brexit Sections', which are then modified by the
presenter for the landing page. Previously we changed the title of the
'brexit section object' in order to benefit from localisations, but this
was overridden by a faulty I18n call in the presenter. This removes the
redundant title change and fixes and re-scopes the localisations to be
specific to the landing page, since they're not used anywhere else.