'Improve' chevron overwrites
The chevron component doesn't handle long text very well. Ideally we need to rewrite the component to make it more resiliant, but this is still in progress and we have had asks to deploy changes sooner.

This adds some additional CSS to make it appear more resiliant. However, it doesn't fix the issue on all screen sizes, and the padding has been removed in some cases to get it to work, so it doesn't look great visually all the time. Hopefully we can remove all of this code soon once the component has been rewritten.