Created Shared translation file
There is a small number of translations that are used in a few different places
across the app so this creates 'shared' translation files to store them.

This also means that since the last of the translations have been moved out of
`config/locales/en.yml` we can delete this file.