Allow scripts to be rendered left to right
If a language should read left to right, this PR allows the shared
`config/locales/en/shared.yml` to be updated with a direction.

When rendering, required divs can use the application helper method
`page_text_direction` to set a dir attribute which will apply
relevent CSS to text.

This follows similar patterns to those in government-frontend and whitehall
https://github.com/alphagov/government-frontend/blob/71927708e97e4d7375ca719cfa861f74e3d7c024/app/helpers/application_helper.rb#L2