Add basic Person page render test
This helps ensure that person pages don't break when govuk_publishing_components
are bumped (for example).

I've had to tweak a few bits of the person model to defend against parts of
the random person schema that is generated.  These may be perfectly valid,
so I don't see it as a bad thing.

The next commit will be adding the schema.org Person schema, so having a
naive working test based on the available govuk_content_schemas seems sensible.