Order High Profile Groups alphabetically
We've had a request from a department to order High Profile Groups
alphabetically on organsation pages. This seems like a reasonable ordering and,
while `ordered_high_profile_groups` on a content item suggests that they should
be ordered, there's no suggestion in the Whitehall codebase that there's any
deliberate ordering or opportunity for users to specify ordering.