Add eCommerce tracking to mainstream browse.
Due to the use of ajax on the page, an initial payload gets sent via a
page view event. All others are sent via an interaction event.

Includes analyitics tests use a find and assert_equal approach, instead
of the has_selector? strategry seen in many of the feature tests in
this repo.

Whilst this introduces an inconsistency with the other tests, it
provides more useful failure state feedback. With has_selector? it can
only ever be true or false. With find and asserts_equal, we get an error
if the element is not present, and otherwise get an expected/actual
comparison.

This is useful here given things like positioning could easily change if
we added further test cases.