Fix wonky paths
We get requests for paths that aren't picked up by standard rails routing, eg:
  https://www.gov.uk/topic%2Flegal-aid-for-providers%2Fmake-application%2Flatest

The slashes in the path are encoded to %2F which means it doesn't match routes
like:

```
get "/topic/:topic_slug/", to: "thing#action"
```
or resources blocks like:
```
resources :topics
```

They seem to get through Router OK ([there's a test exploring this in a branch](https://github.com/alphagov/router/compare/encoded_path_tests?expand=1)),
and are directed to applications that would handle them, should they recognise
the routes.  Government Frontend seems to deal with these OK, because it
uses wildcard routes for the most part, and looks at the document type of the
content item to determine how to deal with it.  Frontend generally works in a
similar way to Government Frontend so is also OK.

Collections relies heavily on knowing URL schemes, such as having a
`/government/organisation` prefix.  This is broken if the URL doesn't
have a slash between `government` and `organisation`.

If the wrong sort of URL falls down to the taxons controller, then we end up
with an error in Sentry (https://sentry.io/organizations/govuk/issues/1399727649/?project=202213&query=is%3Aunresolved).

There a 6000ish errors tracked at present, though we seem to fall back to the
mirrors OK for each of these users are shown stale pages which is not ideal.

I'm not sure where these requests are coming from, but they are for all sorts
of content across GOV.UK.