Move cv route to remove content store call
The DocumentTypeRoutingConstraint used for step by steps makes a call to
content store to verify whether the path refers to a step by step page.

This is fine for general use, but for this page, we want to remove the
requirement for a content store request until we have a content item.