Fix focus behaviour for the links
This removes the components to fix the focus behaviour for the links.
An anchor tag shouldn't wrap a header (which is a block-level tag).

Using the heading component and wrapping it inside an anchor tag
resulted in having the h3 inside the anchor tag, and this made the
link background not behaving as expected when clicked.

Co-authored-by: Christopher Baines <christopher.baines@digital.cabinet-office.gov.uk>