Adds SpecialAnnouncement schema
This adds https://schema.org/SpecialAnnouncement to the landing page.

This should complement the FAQPage schema that is ongoing.

The category field is recommended to be either the wikipedia or wikidata
URLs for this coronavirus 2019 pandemic, so I've picked the wikidata version.

Corresponds to https://github.com/alphagov/govuk-coronavirus-content/pull/21
though it will not break if deployed before that is published.