Add beginning of business page
The skeleton of the business page is based on the coronavirus landing page.

This will be iterated tomorrow.

Note that the `stay_at_home` key in the content item is `header_section`
in the new business content item and there are differences between them.

We'll have an alignment session after this is built.