Redirect coronavirus taxon requests
There is the beginnings of a hierarchy of pages:

/coronavirus (the "landing" page)
  |__ /coronavirus/business-support ("hub" page)
  |__ /coronavirus/education-and-childcare ("hub" page)

We are soon to publish a new branch of the taxonomy which this page hierarchy
will overlay onto. The taxonomy will be used to power navigation, search
and notifications, but we want to use the landing/hub pages as UI for exploring
them, as they're more heavily curated that a taxon page can be.

This commit routes all requests for pages under and including /coronavirus-taxon
and works out where best to send them.

Taxon pages only have a two layer path, so even if a taxon is five levels deep,
it'll still have a path that looks like /sport/badminton-for-pigeon-fanciers.

This means we have to examine the content_item to see if it has an ancestor
that matches one of our hub pages. (Or it might be one of those taxons that
has a hub page itself!)

If it does, then we redirect it to the hub page. If not, we want to send it
to the coronavirus landing page.