Ensure roles match the section they are shown in
This is a slightly ugly fix for a wider problem which is that rather
than linking directly to the role appointments, we're linking to the
people directly. This means that in the link expansion, it's possible
that we'll end up seeing roles that aren't relevant for this particular
section.

For example, if someone has a board member and a special representative
role, they will be linked to twice (`ordered_board_members` and
`ordered_special_representative`) and both roles will appear in both of
those expanded links, so we need to pull out the correct one.

There is a nice fix here to link to the role appointments, but that is
a bigger piece of work so I'll write that up as a separate card. This is
a quick fix to get us to a better point than we were before.