Filter roles to the organisation
This relates to the same problem that was fixed in
https://github.com/alphagov/collections/pull/1718. Because we're linking
to the people, rather than to the role appointments directly, it's
possible that link expansion will include information about roles not
associated with the organisation. We need to filter those out by using
the `ordered_roles` link to ensure the role belongs to the organisation.