Revert #1733 - Featured links automatically showing description
This reverts https://github.com/alphagov/collections/pull/1733

We need to do a deploy of Collections now (6pm on a Friday) to get a
very small link change. I was worried about deploying this also at this
time as this represented a much bigger change to the Coronavirus page
that I don't think anyone is expecting to be shipped for this weekend.

I was also concerned about the usage of `ContentItem.find!` were this
method to fail in a HTTP request it would take down this page. For
example, if one of the links was to start 404ing this whole page would
404. We'd also get some undesirable outcomes if any time of absolute URL
was used (either internal or external).