Show a description on featured links
We're doing this from the content item initially to validate the need.

If we want to do this from content items, we should use links in the
/coronavirus content item if we can, rather than requesting lots of extra
things from the API. We need to make sure that this page can cope with load.

Relies on https://github.com/alphagov/govuk-coronavirus-content/pull/240
being published before they will show.