Pass view into object for context helpers
This model is just used as a means to share presentational logic.
Previously to this it relied on direct access to a view helper to
specify the asset_url. This wasn't contextual and thus relied on
the asset host being absolute for this to produce a full URL.

We can no longer rely on this technique now that we aren't always
setting an asset host, thus we can pass in the view to use the request
context the generate the URL.