Add canonical URL to landing pages
We get this by default on pages that use the machine_readable component,
but we've got a custom set up on these pages that doesn't lend itself to that.

We need to add this because these pages are often linked to with campaign or
marketing query strings. This causes Google et al to believe that they might
behave differently at times. (They don't)