Add Transition landing page AB test
We want to test whether the mention of email signups is putting people off
the checker.

This is only on the `/transition` page, and not `/transition.cy` for now. The
proportion of traffic on the Welsh version is not sufficient to produce quick
statistically significant results.

We add a Z variant to capture traffic that has not been assigned an A/B bucket
from Fastly. This will be shown the standard text, but will not be included in
the test results.