Add lang to person data  object
There are instances on the site where a page will be in Welsh, but
certain components on the page will display data in English as there
is no Welsh translation for those parts of the page.

One of those instances is on the organisations pages rendered by
the collections project. Some of the organisations pages will include a
list of people/ministers which often are not translated. So even when
the document itself is in Welsh, some of the ministers information is
still in English (and not indicated as such). This is a WCAG fail.

Add lang to the person object in people presenter. This supports the
work to fix mixed languages on people/minister information on the
organisation pages. Lang will now be included with the object we are
passing to the image_card component, which renders the markup for each
person in the people lists.