Small corrections based on PR feedback
Refactor contacts_presenter to use presence method instead of empty
string fallback.

Force organisation contact email locale to en. The reason for this is
that this part of the code pulls translations from the application
locale config. However the containing FOI contact block pulls
localisation/translations from Whitehall and is marked as such.

Therefore we could have a scenario where:
1. The page is in a language (usually Welsh)
2. The containing FOI contact block falls back to the English version
   because there is no Welsh translation for those contact details
3. The org contact email is in Welsh again, because at the application
   level there is a Welsh translation for that string. Even though this
would match the language of the page, it still is a WCAG fail as the
containing block is marked with lang="en" (see 2)