Add lang to main element in application template
The <div id="wrapper"> element is a container for the application
breadcrumbs, for the feedback component, and for the <main id="content">
element.

Out of these three things, only the <main> element contains content that
might be translated into languages other than English. Seeing as at this
time both the breadcrumbs and the feedback component are in English, it
makes more sense to add a lang attribute to the <main> element instead
of the wrapper element.

This removes the need to override lang for feedback/breadcrumbs when a
page is being viewed in a language other than English.