Organisations page hidden headings
The bold text in the hidden sections on the government organisations
page looks like a heading but is not one. The meaning of the content
below is tied to the bold bit of text. Therefore it makes sense to use
a heading tag in this context.

Use the appropriate heading tags so that the semantics of this content
matches the visual presentation.

Also use heading tags for the organisation titles/logos in order to
represent content hierarchy accurately.