Load ActiveModel on Coronavirus landing page
When run locally, the coronavirus landing page errored with:

`uninitialized constant #<Class:0x0000562f92d53730>::ActiveModel`

This was being caused by the following line in one of the live stream
views:

`<% if ActiveModel::Type::Boolean.new.cast(live_stream["show_live_stream"]) %>`

This line was added to make sure that the live stream would be correctly
toggled on of even if the value of `show_live_stream` was set to
`"false"` rather than `false`.

The error doesn't occur in production because the dependencies are
front-loaded, but that's not the case locally.