Third level browse links underlines
The "third level" browse links on pages such as this one can only be
distinguished through colour: www.gov.uk/browse/benefits/entitlement

This is a problem as people with certain visual impairments could not
distinguish between what is a link and what is regular text.

This removes the "text-decoration: none" rule that removed the
underlines from the links.