Include missing chief scientific advisor title for board member
The role of chief_scientific_advisor was not being shown when the person
was included on the organisation page as a board member.

This is because each person was expected to have a role of board_member
in order to be displayed.

This is a quick fix where we add the chief_scientific_advisor role as
one of the expected document types returned for board_members.