Style new countdown component
This change also takes the traffic light component into a seperate partial and moves the countdown component from the generic component folder to the transition folder as this is specific to the transition period.

Add support for showing and hiding the countdown

This change also brings the countdown inline with the header

Add support for tablets

Previously the component would not work well for tablets. Now it has a new hybrid style that works well on all screens