Handle invalid postcode error from Mapit
If Mapit is passed an invalid postcode, gds-api-adapters will
return a "400". We need to handle that gracefully so that the frontend
can still show the appropriate "National" Covid-19 restrictions guidance
to users without the page erroring.

Also adds a method to make it easy for the caller to determine if there
has been an error.