Add model for local restriction data
This is a port of a previous commit to finder frontend - https://github.com/alphagov/finder-frontend/pull/2222

It is designed to provide a way to query the data added in https://github.com/alphagov/collections/pull/1934. There are some changes between the PR in finder-frontend and here because of the differences between rspec and minitest. The biggest difference is that minitest lacks the native `stub_const` method that rspec has. This means that I've had to create a new method that returns the file name that is easily stubbable.