Add /results post route
Adds skip_before_action :verify_authenticity_token.
This was because the following error was thrown:

Can't verify CSRF token authenticity.

As we are not submitting and sensitive data, and are just redirecting users,
we don't need this protection on the form.

We don't currently have a session store due to the layers of caching, so we
did not try to set the authenticity token, and have instead disabled it.

Co-authored-by: Leena Gupte <leena.gupte@digital.cabinet-office.gov.uk>
Co-authored-by: Alex Newton <alex.newton@digital.cabinet-office.gov.uk>
Co-authored-by: Jonathan Hallam <jonathan.hallam@digital.cabinet-office.gov.uk>
Co-authored-by: Rosa Fox <rosa.fox@digital.cabinet-office.gov.uk>
Co-authored-by: Dilwoar Hussain <dilwoar.hussain@digital.cabinet-office.gov.uk>