Remove clock component
As we're not using this at present remove it from the codebase. As
this is a distinct commit we can reintroduce it easily enough should we wish.

This component was used in the previous commit which might be useful if we
want to put it back, though it's likely that we need something slightly different
to address the shortcomings revealed in performance analysis.