Test new title and bucket order on transition page
We want to test whether prioritising businesses in the buckets section on
the landing page improves outcomes for all users.

Research shows that people get distracted by the "things you can do now"
section, when they should probably be doing the checker.

As part of this commit I've removed the tests for the old AB variants which
are no longer needed.