Start new transition AB test
We're changing the display of the "things you can do now" links to play
them down a bit. It's a tricky balance: making sure people know they're
there, but also doing the checker as a higher priority.

This replaces the bucket section (which is quite large, bold and contains
multiple sections) with a simple list of links.

This test is English only, so the Welsh page should be unaffected.

https://trello.com/c/3XR9BeaI/511-ab-test-for-urgency-phase-3