Show error if no params sent to postcode checker
The postcode checker has thrown this error a few times:

```
NoMethodError
undefined method `gsub' for nil:NilClass
```

on line:

```
@postcode = params["postcode-lookup"].gsub(/\s+/, "").upcase
```

It looks as though in a handful of cases the form is being submitted without any params.
We should catch this and return the user to the lookup page, rather than throwing a 500 error.

See error in sentry:
https://sentry.io/organizations/govuk/issues/1954271093/events/865d5d86244240ec94e43caf60be4ce8/?project=202213