Adjust "See latest changes" link spacing
On topics pages where a "see latest changes to this content" link exists
as part of the header, there isn't enough spacing between this link and
the main content of the page on mobile.

This adds a margin bottom to the container of this element which helps
space things out for users viewing this page on small devices.