Ensure local restriction hash is empty in tests
The intent of this test is to ensure we return `nil` values for our
properties when the gss code does not exist in our data structure.
However as they stand, the test will pass if the gss code exists, but
has no "name" field.

For example, a YAML structure such as:

```
---
fake code:
  alert_level: 3
```

would still pass the test.

I think what we are actually trying to ensure here is that we create an
empty hash when calling `LocalRestriction#restriction` with a gss code
that is not contained in the YAML/data source.

Asserting that `#restriction` is empty tests this, leaving in the
`assert_nil` on `#area_name` gives us an extra level of confidence that
we have instantiated an empty hash (as opposed to something like an
empty array), though we could arguably achieve this by testing the type instead.

I initially thought `#restriction` might be better off as a private
method, and to me it feels a bit clumsy to reach into the class and
query it directly. I did consider a refactor and exposing some
something like a `#where` or `#empty?` directly on the class, but I
noticed the initial implementation in finder frontend some discussion
was had around this [potentially needing to be
public](https://github.com/alphagov/finder-frontend/pull/2222#discussion_r498826951),
so I've left it for now.

I've also renamed the local variable `restriction` to
`local_restriction` to provide an extra bit of clarity and because
`restriction.restriction` just looked wrong.