Set time zone to London
Borrowed from: https://github.com/alphagov/content-publisher/commit/4ae1c4ed683bf0b24eb11d135aebd3cdd1a8ffa6

We need to make sure the alert level changes go out at the correct time.
The time zone on the server is incorrect, which means that when we're using
dates and times to set the level they're an hour out of date during daylight saving
time. This sets the Local restrictions time zone to 'London'.
We tried doing this at the app level, but it broke a lot of unrelated tests.