Update Atom Feed presenter
- Collections builds the organisation and global atom feeds. It fetches the search
response, and formats the public_timestamp to create an ID. To ensure we don't break
existing feeds by giving old items a new ID, Feed Entry Presenter needs to parse public_timestamp
in the UTC timezone.

- The Organisation Document Presenter uses Date.parse to present the public_updated_at
time stamp, but the test was using Time.zone.parse. This was passing when the default time zone was UTC,
but now Time.zone is set, the two methods return a different string.