Update tests to use fixture data
At the moment we're using the data in the live files for our tests and
that data is constantly subject to change. If we don't change to using
stubbed data, it means that we might need to update the tests when areas
change restriction levels. This makes it harder for non-tech people to
update the levels.

The test data has also been updated to not knowingly use existing gss
codes, so that it's clear is this is test data.