Use action links to future COVID alert guidance
This updates the links to future alert level guidance to be action links.

The locale file has been restructured to support this change.
As we need to be able to get the link text link url separately for use in the action_link component,
it has been possible to DRY-up the config at little and re-use the guidance links from the current
level config.