Use Local Authority name from Mapit
When the restrictions data in "lib/local_restrictions/local-restrictions.yaml" is updated,
a developer is calling Mapit manually to get the name of the local authority to populate
the "name" field.

This seems pointless as we are already get the local authority name
when we do the lookup:

https://github.com/alphagov/collections/blob/master/app/services/location_lookup_service.rb#L37-L40

Asking people to call mapit like this, makes the restrictions data harder to update. Instead we should use
the "name" field in the YAML file as something human-readable for the person updating to look as gss codes are
hard to read, but not something that is used in the code.