Stub document types together
These are used to protect our tests from the govuk_document_types gem
changing. (It's a look up of which document types belong to which super-type).

Our code doesn't really care about it too much, because the looked up document
types are used to query search. It would be a faff to have to change webmocks
if the gem ever changed though, so that's why the stub is there.