Allow override of default test configuration
This change makes it really easy to see what the difference between the
two tests are.

I removed the `send` call to dynamically configure the search-api call. It's
OK, but it does make it really hard to `CMD + F` your way around the code
base. I'll try and clean that up at some point, but those methods are used
in a number of other places in the app, so might be too much for now.