Change content on the Enter Postcode box
This change should make it clearer to users what is required of them
for the postcode input by making the heading and hints less ambiguous.
It also adds a h2 tag to the input label to improve the accessibility
for users of screen readers.

A css override is also necessary as we want to render the hint text
with govspeak but retain the grey colour