Adjust organisation "Follow us" links hidden text
Amends the hidden text for the organisation social media links. The
hidden text will now be empty when the link contains the words "Sign
up". An example of this is one organisation page has a "Sign up for
e-bulletin" link which a screen reader would read as "Follow on sign up
for e-bulletin" which obviously makes no sense.