Only show one lot of alert level text
This is a bit confusing because we've changed the text for the "current"
level, but really all of the changes are in the future.

When we get the data we will need to change the logic in the `results`
view so that it checks for the presence of a future restriction rather
than the "current" one. I have changed this logic yet because I don't
want to make guesses about what the data is going to look like and end
up duplicating work.

As mentioned in a previous commit, I've left the view logic in as we're
going to need to unpick this in a matter of days, and I want to make
that as easy as possible.