Show correct results page for future restrictions
Show the correct results page when there are only future restrictions
for an area.

On 2 December all of England comes out of national restrictions and will
enter local tier based restrictions again.

That means that when the checker goes live again on the 26 November, all
areas in England will only have future restrictions.

The existing logic on the results page ignored future restrictions and
used current restrictions to determine which tier page to show. If we
don't add this logic change then between now and 2 December all areas
would only show the results page for tier 1.

To make this logic a little easier to read and test it has been added to
the model rather than the view.

N.B. The entry for Naboo in the fixture file had to be updated to get
the existing integration tests to pass.

These changes should mean that the page the user sees on 1 December is
the same as the one they are shown on 2 December.

Assumptions
-------------

When the local restrictions data is updated with the new tier
information, all of the existing data will be overwritten.